refactor(frontend): Move search utility functions from types to utils
- Move buildOperatorItems() and buildNullableOperatorItems() from type/search.ts to utils/search.ts
- Move private helper getOperatorLabel() function to utils/search.ts
- Update import statement in AppSearchRuleFilterGroup.vue to use new utils location
- Clean up type/search.ts to contain only types and interfaces
- Improve code organization by separating utility functions from type definitions
