+++ apps/frontend/src/type/search.ts
@@ -1,11 +1,4 @@
-import {
-  type FilterType,
-  type Rule,
-  type RuleGroup,
-  type RuleOperator,
-  nullableOperators,
-  operatorsByTypeMap,
-} from '@beabee/beabee-common';
+import type { Rule, RuleGroup } from '@beabee/beabee-common';
 
 import type { Component } from 'vue';
 
@@ -86,44 +79,4 @@ export interface OperatorLabels {
   date: Record<string, string>;
   array: Record<string, string>;
   all: Record<string, string>;
-}
-
-// Helper function to build operator items from labels
-export function buildOperatorItems(labels: OperatorLabels) {
-  return Object.fromEntries(
-    Object.entries(operatorsByTypeMap).map(([type, typeOperators]) => [
-      type,
-      Object.entries(typeOperators).map(([operator]) => ({
-        id: operator as RuleOperator,
-        label: getOperatorLabel(
-          labels,
-          type as FilterType,
-          operator as RuleOperator
-        ),
-      })),
-    ])
-  );
-}
-
-export function buildNullableOperatorItems(labels: OperatorLabels) {
-  return Object.entries(nullableOperators).map(([operator]) => ({
-    id: operator as RuleOperator,
-    label: getOperatorLabel(labels, 'all', operator as RuleOperator),
-  }));
-}
-
-function getOperatorLabel(
-  labels: OperatorLabels,
-  type: FilterType | 'all',
-  operator: RuleOperator
-): string {
-  let labelType = type;
-  if (operator === 'is_empty' || operator === 'is_not_empty') {
-    labelType = 'all';
-  }
-  if (labelType === 'contact') {
-    labelType = 'text';
-  }
-
-  return labels[labelType]?.[operator] || operator;
 }
